+++ @ecomplus/storefront-app/src/components/EcAccountForm.vue
@@ -1,3 +1,0 @@
-<script src="./js/EcAccountForm.js"></script>
-<template lang="html" src="./html/EcAccountForm.html"></template>
-<style lang="scss" src="./scss/EcAccountForm.scss"></style>