+++ @ecomplus/storefront-components/src/AccountForm.vue
@@ -1,0 +1,2 @@
+<script src="./js/AccountForm.js"></script>
+<template lang="html" src="./html/AccountForm.html"></template>