+++ @ecomplus/widget-compre-confie/cms.config.js
@@ -1,0 +1,53 @@
+import { name } from './package.json'
+
+export default () => ({
+  label: 'Avaliações Compre & Confie',
+  fields: [
+    {
+      name: 'pkg',
+      widget: 'hidden',
+      default: name
+    },
+    {
+      label: 'Ativar',
+      hint: 'Instalar selo Compre & Confie',
+      name: 'active',
+      widget: 'boolean'
+    },
+    {
+      label: 'Apenas em desktops',
+      name: 'desktopOnly',
+      hint: 'Desativa o widget em dispositivos móveis',
+      widget: 'boolean'
+    },
+    {
+      name: 'enableCheckout',
+      widget: 'hidden',
+      default: true
+    },
+    {
+      name: 'disablePages',
+      widget: 'hidden',
+      default: true
+    },
+    {
+      label: 'Opções',
+      name: 'options',
+      widget: 'object',
+      hint: 'Personalizações do widget',
+      fields: [
+        {
+          label: 'Store Id Compre & Confie',
+          name: 'compreConfieStoreId',
+          widget: 'string',
+          hint: 'ID da loja que é informado pelo Compre & Confie'
+        }
+      ]
+    },
+    {
+      name: 'stampsAppend',
+      widget: 'hidden',
+      default: 'src/append/stamps'
+    }
+  ]
+})