+++ @ecomplus/storefront-template/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.45](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.44...@ecomplus/storefront-template@2.0.0-beta.45) (2020-07-03)
+
+
+### Features
+
+* **compre-config:** compre & confie storefront widget ([#258](https://github.com/ecomplus/storefront/issues/258)) ([0400e40](https://github.com/ecomplus/storefront/commit/0400e40231604a5ffc8ae22fefa67ae3cc034bae))
+
+
+
+
+
 # [2.0.0-beta.44](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.43...@ecomplus/storefront-template@2.0.0-beta.44) (2020-07-02)